==> 1_contextual-filtering.ipynb <==
import pandas as pd

# Load file with available points

# the city information is available in the `simplemaps.csv` file
df_all = pd.read_csv('data/global-streetscapes/data/simplemaps.csv', index_col=0)
df_all.head()

# Filter based on cities

# Singapore, Amsterdam, San Francisco, Abuja, Santiago 
city_id = [1702341327, 1528355309, 1840021543, 1566342270, 1152554349]

df_city = df_all[df_all['city_id'].isin(city_id)]
df_city.groupby("city").size()

# Filter contextual information

The filtering criteria is:

- No panoramic images: `pano_status` == False
- Good quality `quality` == good
- No reflection `reflection` == no
- During the day: `lighting_condition` == day
- Clear weather :`weather` == clear
- View direction `view_direction` == `front/back`

# load contextual information
df_contextual = pd.read_csv("data/global-streetscapes/data/contextual.csv")
df_season = pd.read_csv("data/global-streetscapes/data/season.csv")


# merge our filtered dataset with contextual data
df_city_merged = df_city.merge(df_contextual, on=['uuid', 'source', 'orig_id'])
df_city.groupby("city").size()

# No panoramic

df_filtered = df_city_merged[df_city_merged['pano_status'] == False]
df_filtered.groupby("city").size()

# Quality

df_filtered = df_filtered[df_filtered['quality'] == 'good']
df_filtered.groupby("city").size()

# Reflection

df_filtered = df_filtered[df_filtered['reflection'] == 'no']
df_filtered.groupby("city").size()

## Lighting condition

df_filtered = df_filtered[df_filtered['lighting_condition'] == 'day']
df_filtered.groupby("city").size()

# Weather

df_filtered = df_filtered[df_filtered['weather'] == 'clear']
df_filtered.groupby("city").size()

## View direction

df_filtered = df_filtered[df_filtered['view_direction'] == 'front/back']
df_filtered.groupby("city").size()

# Save to csv 

df_filtered.to_csv('download_imgs/selected_points.csv')
