==> 1b_visual-complexity-analysis.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from scipy.stats import entropy

# Load filtered file and file with segmentation results

# load filtered points from all countries
df_filtered_pts = pd.read_csv("download_imgs/selected_points.csv")
df_filtered_pts.groupby("city").size()

# load filtered points from abuja
df_filtered_pts_abuja = pd.read_csv("download_imgs/selected_points_abuja.csv")
df_filtered_pts.shape[0]

# The global streetscapes have the results of semantic segmentation
df_seg = pd.read_csv(
    "data/global-streetscapes/data/segmentation.csv"
)
df_seg

# The global streetscapes have the results of semantic segmentation
df_seg_abuja = pd.read_csv("data/abuja/abuja_segmentation.csv")
df_seg_abuja

# Merge filtered points and segmentation values

df_filtered_pts = pd.merge(df_filtered_pts, df_seg, on=["uuid"], how="left")
df_filtered_pts

df_filtered_pts_abuja = pd.merge(df_filtered_pts_abuja, df_seg_abuja, on=["uuid"], how="left")
df_filtered_pts_abuja

# Visual complexity calculation

# remove images where segmentation failed
print("Images with 0 value in total pixel count are a result of failed segmentation")
print(
    f"Num. images where segmentation failed: {df_filtered_pts[df_filtered_pts['Total'] == 0].shape[0]}"
)

df_filtered_pts = df_filtered_pts[df_filtered_pts["Total"] != 0]
df_filtered_pts.shape

# remove images where segmentation failed
print("Images with 0 value in total pixel count are a result of failed segmentation")
print(
    f"Num. images where segmentation failed: {df_filtered_pts_abuja[df_filtered_pts_abuja['Total'] == 0].shape[0]}"
)

df_filtered_pts_abuja = df_filtered_pts_abuja[df_filtered_pts_abuja["Total"] != 0]
df_filtered_pts_abuja.shape

idx_class_1 = df_filtered_pts.columns.get_loc("Bird")
idx_class_1
# Loop through semantic classes (mapillary has 65 classes) and compute Visual Complexity
df_indicators = df_filtered_pts.iloc[:, idx_class_1:]
list_indicators = df_indicators.to_dict("records")

# Keep track of calculations
numpy_array_list = list()

# All_indicators is a dictionary containing the number of pixels per semantic class for each image
for all_indicators in list_indicators:
    temp = list()

    for s_class_key, s_class_val in all_indicators.items():
        if s_class_key == "Total":
            break  # only calculate ratio for the 65 classes
        else:
            # all_indicators['Total'] contains total number of pixels of an image
            prob = round(s_class_val / all_indicators["Total"], 4)
            temp.append(prob)

    numpy_array_list.append(np.array(temp))

# Add Visual Complexity and remove images with threshold less than <1.0> (probably pointing at road or wall)
df_filtered_pts["visual_complexity"] = [entropy(array) for array in numpy_array_list]

idx_class_1 = df_filtered_pts_abuja.columns.get_loc("Bird")
idx_class_1
# Loop through semantic classes (mapillary has 65 classes) and compute Visual Complexity
df_indicators = df_filtered_pts_abuja.iloc[:, idx_class_1:]
list_indicators = df_indicators.to_dict("records")

# Keep track of calculations
numpy_array_list = list()

# All_indicators is a dictionary containing the number of pixels per semantic class for each image
for all_indicators in list_indicators:
    temp = list()

    for s_class_key, s_class_val in all_indicators.items():
        if s_class_key == "Total":
            break  # only calculate ratio for the 65 classes
        else:
            # all_indicators['Total'] contains total number of pixels of an image
            prob = round(s_class_val / all_indicators["Total"], 4)
            temp.append(prob)

    numpy_array_list.append(np.array(temp))

# Add Visual Complexity and remove images with threshold less than <1.0> (probably pointing at road or wall)
df_filtered_pts_abuja["visual_complexity"] = [entropy(array) for array in numpy_array_list]

df_filtered_pts_abuja['city'] = 'Abuja'
df_filtered_pts_abuja

# Visual complexity histograms

df_filtered_pts = df_filtered_pts[df_filtered_pts['city'] != 'Abuja']
df_filtered_pts_all = pd.concat([df_filtered_pts[['uuid', 'city', 'visual_complexity']], df_filtered_pts_abuja[['uuid', 'city', 'visual_complexity']]])

df_filtered_pts_all['city'].unique()
df_filtered_pts_all.to_csv('data/svi/visual_complexity_all.csv')
