==> 3_random-sample.ipynb <==
import math

import numpy as np
import pandas as pd

# Load file with clustered data

df_clustered = pd.read_csv("download_imgs/clustered_points.csv")
df_clustered

df_clustered.groupby("city").size()

# Random sample

# for each city, randomly sample a total of 100 images, evenly distributed from each cluster
city_id = [1702341327, 1528355309, 1840021543, 1566342270, 1152554349]

city_names = {
    1702341327: "Singapore",
    1528355309: "Amsterdam",
    1840021543: "San Francisco",
    1566342270: "Abuja",
    1152554349: "Santiago",
}


df_samples = pd.DataFrame(columns=df_clustered.columns)

NUM_SVIS = 100

df_clustered.groupby("city").size()

# keep track of the samples
df_sampled = pd.DataFrame()

# keep track of already sampled indices
sampled_indices = []

# iterate over each city
for id in city_id:

    # DEBUG
    if id == 1566342270: # skip Abuja until the new images are downloaded.
        continue
    # DEBUG
    
    print(f"City: {city_names[id]}")
    df_city = df_clustered[df_clustered["city_id"] == id]

    # count the number of rows in each cluster
    cluster_counts = df_city["cluster_label"].value_counts()

    # calculate the number of samples to take from each cluster
    samples_per_cluster = np.floor_divide(NUM_SVIS, len(cluster_counts))

    # keep track of the samples
    df_sampled_city = pd.DataFrame()

    # sample from each cluster
    for cluster_label, count in cluster_counts.items():
        if count <= samples_per_cluster:
            # if the cluster has fewer samples than required, take all samples
            df_sampled_city = pd.concat(
                [df_sampled_city, df_city[df_city["cluster_label"] == cluster_label]]
            )
        else:
            # sample evenly from the cluster
            available_samples = df_city[df_city["cluster_label"] == cluster_label].index.difference(sampled_indices)
            if len(available_samples) == 0:
                continue  # Skip this cluster if no available samples
            sampled_from_cluster = np.random.choice(available_samples, size=samples_per_cluster, replace=False)
            sampled_indices.extend(sampled_from_cluster)
            df_sampled_city = pd.concat(
                [
                    df_sampled_city,
                    df_city.loc[sampled_from_cluster],
                ]
            )

    # if the total number of samples taken so far is less than 100, take additional samples
    remaining_samples = NUM_SVIS - len(df_sampled_city)

    if remaining_samples > 0:
        print(f"There are still {remaining_samples} sample(s) required")
        for cluster_label, count in cluster_counts.items():
            if count > samples_per_cluster:
                # sample from clusters with more samples
                available_samples = df_city[df_city["cluster_label"] == cluster_label].index.difference(sampled_indices)
                if len(available_samples) == 0:
                    continue  # Skip this cluster if no available samples
                additional_samples = min(remaining_samples, len(available_samples))
                new_samples = np.random.choice(available_samples, size=additional_samples, replace=False)

                sampled_indices.extend(new_samples)
                df_sampled_city = pd.concat(
                    [
                        df_sampled_city,
                        df_city.loc[new_samples],
                    ]
                )
                remaining_samples -= additional_samples
                if remaining_samples == 0:
                    break

    # ensure that the final sampled DataFrame has exactly NUM_SVIS rows
    df_sampled_city = df_sampled_city.sample(n=NUM_SVIS, random_state=42)

    # reset index
    df_sampled_city.reset_index(drop=True, inplace=True)

    # concatenate all cities
    df_sampled = pd.concat([df_sampled, df_sampled_city], ignore_index=True)


df_sampled

# Save to csv

df_sampled["uuid"] = df_sampled["uuid_x"]
df_sampled.to_csv("download_imgs/filtered_points.csv")
df_sampled

df_simple_metadata = df_sampled[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]
df_simple_metadata.to_csv("download_imgs/metadata_points.csv")
df_simple_metadata

df_sampled["uuid"].nunique()

df_sampled.groupby("city").size()
