==> 4_resample_missing_Abuja.ipynb <==
# Steps to add new samples
1. Run entire pipeline and save images in `v1/`
2. Manually go through images and write down the name (`uuid`) of the images that are not suitable.
3. Remove this images and save the resulting ones in `v2/`
4. Launch this notebook to download missing samples. Repeat as needed

import numpy as np
import pandas as pd

df_clustered = pd.read_csv("download_imgs/clustered_points_abuja.csv")
df_clustered

df_metadata_downloaded = pd.read_csv("data/svi/sampled/metadata_points_abuja.csv")
df_metadata_downloaded

# Cities

city_names = {
    1702341327: "Singapore",
    1528355309: "Amsterdam",
    1840021543: "San Francisco",
    1566342270: "Abuja",
    1152554349: "Santiago",
}

# Amsterdam

## First pass

REMOVED_SVIS_1 = [
    # first pass
    "0fb1e7d4-e90f-4de5-8b10-1f281105d244",
    "1e7d1c74-d301-4c1d-8776-c88ee8acd0ff",
    "1f864f61-9715-46f7-a067-43f0eba476b4",
    "2ae38d4a-01e7-4183-9618-b87c25d6f8a7",
    "3fa9e9c3-9a75-4c85-bb7d-d5b86e011aa7",
    "4fba109b-7dba-4ec4-965f-481f1db87b9c",
    "6c4b7c9d-ba52-4958-8237-1dce07fb1139",
    "6d715c5c-5035-48e8-9870-d51bafc7c4e2",
    "6d5398dd-bd1d-4e90-aa2b-21500b660e17",
    "8c60b85f-ff46-40ee-b861-d6f260a5176b",
    "8c415faa-4031-42a5-8a2d-12a0434a4e33",
    "7fbd79ea-7c29-47d4-b9a2-e13c83db9c04",
    "8d8bad30-8ada-4556-b223-b0387c075517",
    "9f51ee1c-9729-4de6-9af2-261f092dcff5",
    "7bf7a5b4-d435-4509-acf3-6895a8d19c93",
    "9fb96770-4848-426b-bc21-e5da4dd71626",
    "10c03ca2-5014-4aa6-8cd6-0771ae435627",
    "38d3db02-a693-4aa9-b152-291231df0b5e",
    "57b4900c-d933-40be-9ed2-f6427ed769ac",
    "70bccbf5-f658-4113-93f6-0dcf387a92e9",
    "81e94aea-e464-473d-b88b-e914a5a4444c",
    "94f104b9-69bb-4756-82cd-a9badcfb273c",
    "96cedc52-2a18-42e9-848a-3cf6722a4766",
    "0104ce21-a6e6-4432-a7c8-33955e1e1dd1",
    "241d0bf3-83fd-44ab-a8e0-d1ec9acfa65e",
    "301b7a1f-51cb-4cbd-8542-a329a90d190d",
    "701ecd08-2e29-4db3-9bae-3211935d1dc9",
    "902d123c-9d84-46d4-95b0-01d159f730c1",
    "4740b91b-23fc-45ad-bd07-24d5d4c68214",
    "5955ad5e-1daa-445f-b88a-18013eb6659f",
    "7331e36d-405b-465f-bdb3-b16851d53b45",
    "9938d890-50a1-488e-9602-606f30322dc1",
    "324672d6-b7cc-476d-8961-339d0aed08aa",
    "328173bd-7d41-4bd7-af9b-801f3b6a33e0",
    "563114ae-f885-40f9-b400-5c79265ee418",
    "912517f5-96de-49fe-9426-f5d19c76a46f",
    "1644163e-2fdd-4c05-953c-1d00cc77f6bc",
    "22775926-8137-40b0-884f-7a5eb17531e3",
    "65747777-a5f5-4981-bc3b-c9524bf17249",
    "a8e3f1e7-39b0-4ebd-8503-fbacdfa345b1",
    "aca6d4af-e61e-499f-a62f-793d1b9b329c",
    "ad5f6835-be55-466a-acab-7af42dc809d1",
    "b5ec19d0-0ab3-4477-9045-bd6236f77325",
    "b7ece355-c436-404f-a132-68ae5569ad5d",
    "b73ac2ef-e8e1-4fba-a062-0224ae3be716",
    "b947d350-5de0-45b6-9410-d4356d2b02d4",
    "bdf8cce8-dfee-465e-b79b-09a4d5777a14",
    "c2b4cbb2-d27e-4b16-9f30-8a8892decaf0",
    "c6c37d98-d00f-4d4f-9c5f-8316b2309489",
    "cc0b90df-5ec6-4d14-a2ac-2a4574a3ac45",
    "cc88b795-f87e-42d6-906b-b73ec1ca80d1",
    "cdc809dd-21b4-4690-afab-7c820b651ee5",
    "d4ec46d9-c90a-43e3-a659-f524b4458dd4",
    "dc0d6463-74c8-4c8b-bf6d-a547d086663f",
    "e38f1d60-a581-42fb-891a-dcfbf2a9e0a7",
    "e46c1d96-e709-4629-9064-bef4ed22b97d",
    "e123d608-71f7-4e44-bd07-468e1a05e28e",
    "ee7e1929-8e64-49d4-9a5b-d57e4d8af61e",
    "ef9ea2ba-d8e7-4f36-820d-75806a67e9a4",
    "ef0466f5-f528-4ac1-957f-4af4411bafa3",
    "f0bf4fbe-63b5-4232-a00f-10e9f5ccd717",
    "f1b6d0f8-4311-46e8-9fb5-dc00f0e0aef8",
    "f856430b-26d2-43e4-aed1-abe307cc062e",
    "f8157329-5333-4031-bb53-8ce254332741",
    "fa0966d5-23f3-41a8-9787-0feeed475307",
    "ff7faefd-7455-4d7a-b7df-e22b148d9c45",
]

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid"].isin(REMOVED_SVIS_1)]
df_removed.groupby("cluster_label").size()

df_city = df_clustered.copy()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid"].isin(df_metadata_downloaded["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city.to_csv("download_imgs/filtered_points_abuja_v2.csv")

## Second pass

REMOVED_SVIS_2 = [  # second pass
    "0a40fa26-a053-47b4-b5f0-954b9d97b106",
    "0cfc766c-4d39-448f-b889-79d58f2048af",
    "1c6ef890-8832-4d4c-871b-cefa7e28f508",
    "2f2e7462-4d4f-49f8-b2c5-a5c83a4c8b73",
    "3d4cb6de-10b9-4ed1-9836-9c9be5cb9440",
    "f7dcc4fa-e984-4a3c-8e9f-c9a35f66841c",
    "f0dc273d-6708-4a6b-a9f8-e5d293d60caf",
    "eb15c621-227a-4a09-81d5-42ebd44c8c97",
    "ea3cb12b-fd2a-419c-8c06-231817f7218b",
    "d809d1ab-93d1-4cc1-b052-d40dd4e4c8c4",
    "d1d3d8cb-1c1e-4680-a6db-4ab03d0bd611",
    "cf631fc6-3185-4828-9898-66f5c4f41eca",
    "ce39d9a4-2a63-4408-9332-52abfd4a42fb",
    "cb123235-ab38-43bb-b39f-b537dc2e62d8",
    "c839a65a-c7fb-4ae7-b8de-4dd0d6714ce4",
    "c4ae6278-3b9c-4ef3-8924-05f4af51c546",
    "ab5a574f-d333-4c44-a900-0d3fdee156a6",
    "a4a3877d-2513-481e-b4a4-c6cbf50d354b",
    "a2e3eda3-58d3-477d-b69d-14d1d416337f",
    "37404431-adaf-4f55-85d4-11d99c6bc1b3",
    "0605466f-8a3d-44b0-9c40-960c55efa112",
    "76002b58-2edd-48bf-9766-3b40e5a75a3f",
    "5efa3b12-8365-46b4-b182-2e975c853e4a",
    "06b0af24-dce9-480c-9c3b-89de37f8bd40",
    "7cf8b2d8-1bce-4c76-b306-10cd5493c893",
    "5c2bebbd-d750-4bcb-9b8a-43b4121e077c",
    "8d668d5b-0478-4922-a25b-cfcf03976dfb",
    "9ab338ae-ebe6-477d-8807-d2e4df85b853",
    "9b8f0e77-4fce-44ea-992a-f80486cd2ce3",
    "37a4db20-93d0-4970-a913-a88fa1638562",
    "48d4cf45-6866-4375-8f7e-b864b56a5ad9",
    "53ead40d-7b05-4a17-a37c-3d903ef4060d",
    "70ebfd09-0093-443a-8678-e24e9ae2ab7d",
    "74c6d404-4e45-45fb-a7f1-2931bb5038e2",
    "00131e0e-2b13-40be-aeaa-876c36c79a53",
    "361d932f-3701-4a50-bddf-aeb4a3c24e77",
    "385eb5b2-60be-4c71-b1c8-1e77de0dc704",
    "0592ef1d-fa39-4e0f-9267-5643ab3361ff",
    "760ff854-aa8c-421b-baff-0d8a2aa7602f",
    "2159a6e9-0663-4455-add4-76deda288bdb",
    "8954aac2-dda5-4421-9396-0a75521e92ba",
    "bcaffe73-1101-48c2-974e-297109f83929",
    "63aff7ac-0a0b-4f37-913f-16fdca36e63a",
]

df_v2_metadata = df_sampled_city[["uuid", "cluster_label"]]

df_metadata_all = pd.concat([df_metadata_downloaded, df_v2_metadata])

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid"].isin(REMOVED_SVIS_2)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid"].isin(df_metadata_all["uuid"]))
    ]

    if len(available_samples) <= count:
        sampled_from_cluster = available_samples.copy()
    else:
        sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city.to_csv("download_imgs/filtered_points_abuja_v3.csv")

## Third pass

REMOVED_SVIS_3 = [
    "1b226569-8747-4439-96cd-311eb62495f1",
    "1c6ef890-8832-4d4c-871b-cefa7e28f508",
    "4be23a74-9a71-4313-b0ee-ad54839304f7",
    "7cf8b2d8-1bce-4c76-b306-10cd5493c893",
    "8d668d5b-0478-4922-a25b-cfcf03976dfb",
    "9c0401c4-7b12-4579-aa49-11025e4f6cd2",
    "48d4cf45-6866-4375-8f7e-b864b56a5ad9",
    "21adc08a-9564-452f-9af1-a746ffa32e00",
    "70ebfd09-0093-443a-8678-e24e9ae2ab7d",
    "74c6d404-4e45-45fb-a7f1-2931bb5038e2",
    "f1dcea74-7242-4b45-8328-1b72a60ffa35",
    "eb15c621-227a-4a09-81d5-42ebd44c8c97",
    "cf631fc6-3185-4828-9898-66f5c4f41eca",
    "ce994bb6-25a1-41eb-ad6e-ca84ebc646c7",
    "cb123235-ab38-43bb-b39f-b537dc2e62d8",
    "b4424dec-8136-4da2-ac50-d3f602b6ebef",
    "b998b3bb-d357-4c39-b37f-cbd586a18674",
    "624f5eb0-156e-43b9-a418-0e687d860974",
    "4779e1c9-fa97-4185-9836-fb77daed0b91",
    "5379bec6-8fa8-4df0-a72b-f0a7380e3d00",
    "6775ec9e-2803-44c1-8b0b-dcf254c22920",
    "78981c15-73e2-44a9-903b-b713bcbfa949",
    "19700864-e052-4790-b1a4-52cd441fc89a",
    "0592ef1d-fa39-4e0f-9267-5643ab3361ff",
]

df_v3_metadata = df_sampled_city[
    ["uuid", "cluster_label"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v3_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid"].isin(REMOVED_SVIS_3)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid"].isin(df_metadata_all["uuid"]))
    ]

    if len(available_samples) <= count:
        sampled_from_cluster = available_samples.copy()
    else:
        sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city.to_csv("download_imgs/filtered_points_abuja_v4.csv")

## Fourth pass


REMOVED_SVIS_4 = [
    "0a673be4-521d-4f26-b384-02ffd85f8851",
    "6afcb7d3-b760-440d-8802-e19744316cd5",
    "47a84fe3-9850-443c-8562-dbfe78354673",
    "78c89be1-3a36-4474-9cac-cbec7f482169",
    "530a4dff-5cc5-4423-9bf5-cec9a07ea809",
    "705d327f-d06a-444c-ba2b-76f48a58b24f",
    "3641afdd-f429-41fe-a368-1cec42e50d6f",
    "aa278d95-78b1-4350-9e04-91bb47981e01",
    "b1ca22f3-2a04-47e6-8d52-4d77e9c69fd0",
    "b9c996ef-ec98-4736-bdc1-a2536192c7a0",
    "d2e0a61b-e580-4089-ba1e-5919f42fb425",
    "d49daaee-eea6-4ef1-9d5a-ed56731dfd6f",
    "d314489b-119f-43d0-b1fa-260a6dbfb17c",
]

df_v4_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v4_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_4)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v4.csv")

## Fifth pass

REMOVED_SVIS_5 = [
    "0e2906d6-e61a-4b08-bc46-4b91b55f8e6a",
    "96ca31a0-9729-45b9-91dc-e138c543566f",
    "339db626-bf2c-4ab4-a3f4-ef7719a3998d",
    "98864503-8496-411f-a894-86042fd0d23b",
    "b12933ce-4105-4d21-b8c5-d9777bcc9d8c",
    "b9360691-eaec-4443-bb2f-f91aac116128",
    "e843b57b-6d17-4512-acbd-1ea868bc14f0",
    "e1805d35-00bd-4161-8272-39f9e7eee751",
]

df_v5_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v5_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_5)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v5.csv")

## Sixth pass

REMOVED_SVIS_6 = [
    "1d41ea4d-7d30-4284-9a29-3f15903afc21",
    "6bb34eb1-adc3-4b99-8b34-993362ab41ae",
    "d630db07-cd16-491e-aa97-fd71b7ce5ea3",
    "f1187b27-5601-49f4-b2a4-16e425bda601",
]

df_v6_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v6_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_5)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v6.csv")
