==> 4_resample_missing_Amsterdam.ipynb <==
# Steps to add new samples
1. Run entire pipeline and save images in `v1/`
2. Manually go through images and write down the name (`uuid`) of the images that are not suitable.
3. Remove this images and save the resulting ones in `v2/`
4. Launch this notebook to download missing samples. Repeat as needed

import os

import numpy as np
import pandas as pd

df_clustered = pd.read_csv("download_imgs/clustered_points.csv")
df_clustered

df_metadata_downloaded = pd.read_csv("data/svi/sampled/metadata_points.csv")
df_metadata_downloaded

# Cities

city_names = {
    1702341327: "Singapore",
    1528355309: "Amsterdam",
    1840021543: "San Francisco",
    1566342270: "Abuja",
    1152554349: "Santiago",
}

# Amsterdam

## First pass

REMOVED_SVIS_1 = [
    # first pass
    "1acd163c-c094-4af1-80b6-b0680db1c175",
    "1c6779f6-193d-427b-bd53-b025a6ee3016",
    "1ef031c6-af43-4d6d-ab95-2fc976102f35",
    "2c3da425-1f25-414c-a8ba-859d7cad2fbd",
    "03cdff0c-aee3-45f7-ad1b-8cab5c249d03",
    "3e94d5ae-3c8a-4b7d-91e3-645e86fd5ccc",
    "4c662032-2137-4343-bdb9-990ba3f98df0",
    "5dab8505-6e9a-4d2b-b186-5cdcd083ef1a",
    "6ad8365c-7f5a-465b-9051-0003a152721b",
    "7c7a938f-294a-432f-9b0b-55a101dc2fbd",
    "7e69bc5b-80d4-4bff-b7cf-706d9dca87ae",
    "8d791276-cbe4-4978-b04f-7ec770fc7cd0",
    "9a86d209-cb1d-498e-afed-5e0e8d9d1653",
    "26a7462b-9c54-405f-a2dd-9bbdbd639eb3",
    "26c48205-c4c3-4356-8431-c956b341bf1b",
    "33b1a727-ef81-45f5-b6b5-080824c0b157",
    "039ad99a-1035-4904-ae3e-562501cb8e0d",
    "59be72e8-831d-4140-8844-6fef0adbf98f",
    "78ae037d-166c-499c-a4b7-fc686e185631",
    "98f53d0c-642b-4d35-9721-544e9723f222",
    "6296ee36-f645-4272-872e-c869927541ce",
    "7877edf2-0bcc-4c00-ac9d-fac2f12c43af",
    "8346a1b2-8c29-491a-9393-bea8c54509de",
    "90579e51-efc7-4a55-bb06-96e79bc1c1e1",
    "235621b8-d4a6-4c14-b2b1-262947b14d1d",
    "705302ab-b473-4d2c-abfb-1c5d76dd3420",
    "845943b6-d8dc-4b98-9c40-43dde4fd72a9",
    "38219373-8b3f-442a-bd0a-dedc65504ae2",
    "a157d168-e868-49db-ab09-221084e1f1e4",
    "a5115ca3-4fb4-44cf-b74b-2983b426c442",
    "a8980cf3-3387-450e-9e87-8bc2e6288de7",
    "addad49b-b43c-4c6d-92c1-f3d76d1ea6c8",
    "ba379356-3227-4c15-94a4-3a8eaad40d4a",
    "bae73196-8f70-427e-8e59-989bb1ae4eee",
    "bb7f67a3-2e41-46cc-8509-66e2f4799beb",
    "c9afb817-4519-4d7a-909e-17ea3cecb031",
    "c506886a-451f-4d42-9cdc-1332d3930392",
    "ca3e74b2-d891-488e-a551-5a5be071f124",
    "de9b9d23-c823-4607-9ac0-96f1adcfbd97",
    "ebd602d7-4962-4b31-b7af-3d3ef19744f6",
    "ed2eb01e-e599-4a40-bc77-74d025852e32",
    "f7a93935-3281-4370-8572-6c1cbef713cf",
    "f9971170-c1a5-450f-bb85-3c7daeb43c1f",
    "fe26594f-52ca-4936-8c81-1e96407312dc",
]

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_1)]
df_removed.groupby("cluster_label").size()

df_city = df_clustered[df_clustered["city_id"] == 1528355309]  # amsterdam

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_downloaded["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam.csv")

## Second pass

REMOVED_SVIS_2 = [  # second pass
    "0cea9994-5621-4c89-9353-a9de7cf1371f",
    "1e4cebae-68c4-4358-9de1-18e5edf104fe",
    "2ca89660-a34c-4194-baa3-35beb370c8ef",
    "4b7dac56-6685-4278-bd9e-f809ea7843a0",
    "4dc319ad-4d09-4528-a7f4-a731c7f71e65",
    "6c302b33-9e3b-46bf-b020-ac09347ea79f",
    "7db2851e-b7b3-4581-9846-8b6c425ddd19",
    "8d55e43c-ed98-4d60-8819-f88e98f073c2",
    "85ad1a6d-654d-4ca1-b616-a80ec5ad4749",
    "93c4e7b3-9af2-4729-b96e-dea9d375ea63",
    "099de9e4-c225-4200-991c-8ec147c7365b",
    "f18457b7-534e-40d9-b03c-197a71deb17d",
    "f17d4a1e-3391-4198-8a46-7ecfdfe73dd0",
    "f6083463-afda-4dbe-b366-7a695db22e59",
    "c414c1e6-eeea-4aaa-8bcc-68ff5aa58905",
    "c73feffc-a856-4561-8ba2-b2e311d00a94",
    "b8f2e47d-e8e8-435d-8722-5c73c43e5d22",
    "b8cf6446-7286-4c6b-883c-cab743d0a2cf",
    "ad82cf79-c169-4557-bcf6-b507d8e3efbf",
    "a74c67d6-aa62-4103-bd2a-d1ad0f362840",
    "a35e8045-7c2a-452b-9b23-6891cee20cf7",
    "72732527-fd3b-48b4-836c-42873e591645",
    "54477402-e17c-4d3c-b89f-fae4ff7e845c",
    "101dd8de-fe13-4c87-9843-53787cb66bf4",
    "0448be45-cfbf-428c-a6d7-31e5180e36a7",
    "507c7788-4a4e-4d94-83ad-8386534b0111",
    "1340ffeb-6eef-4aff-9cd0-ff8f450f82c1",
]

df_v2_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_downloaded, df_v2_metadata])

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_2)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v2.csv")

## Third pass

REMOVED_SVIS_3 = [
    "0b1563a9-1766-4822-b985-9a1d7e83a1bc",
    "0f5f6446-1620-48de-8f6b-6838f934f879",
    "3be8d823-912d-4af7-b301-8b2976c8daa2",
    "3dce7dc2-d350-4287-a645-e2fb3383ac07",
    "3e505a42-e538-4f73-9b8d-4accb37b5550",
    "7c06a2a8-db3e-43a2-aced-d2ad1abf5a54",
    "7de12258-9e2b-4ecf-b34d-dd5712aef0fb",
    "2e3c7515-ee55-40af-9b52-edc34549e391",
    "9c5cc50b-39b5-4a4b-b299-8d1c4ff49703",
    "22bded19-d26b-4d83-9a30-89a06ca767df",
    "674fc498-a8bc-4642-b8a5-7ad53888f8b9",
    "fa4879a1-c2c1-4171-bd83-82a22cdd079b",
    "ef1aa861-cae8-457e-9d55-cbf3e5f2a533",
    "62660ed8-a4b4-4b74-a1d7-5cb1887adeae",
    "0703ee7c-6080-4699-b47e-2290a5ee3949",
    "59d7f57f-6637-4888-8d15-1b88bbb0d3d5",
    "6027e6cf-3b6f-4cc1-a955-3ab20af5ea36",
    "59716a40-e6c9-4597-8193-4d9614244139",
    "a2eae891-1053-4a69-b8dc-73574ed06491",
    "efa998ea-d48a-4c18-b978-1a2870fae4c6",
]

df_v3_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v3_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_3)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v3.csv")

## Fourth pass


REMOVED_SVIS_4 = [
    "0a673be4-521d-4f26-b384-02ffd85f8851",
    "6afcb7d3-b760-440d-8802-e19744316cd5",
    "47a84fe3-9850-443c-8562-dbfe78354673",
    "78c89be1-3a36-4474-9cac-cbec7f482169",
    "530a4dff-5cc5-4423-9bf5-cec9a07ea809",
    "705d327f-d06a-444c-ba2b-76f48a58b24f",
    "3641afdd-f429-41fe-a368-1cec42e50d6f",
    "aa278d95-78b1-4350-9e04-91bb47981e01",
    "b1ca22f3-2a04-47e6-8d52-4d77e9c69fd0",
    "b9c996ef-ec98-4736-bdc1-a2536192c7a0",
    "d2e0a61b-e580-4089-ba1e-5919f42fb425",
    "d49daaee-eea6-4ef1-9d5a-ed56731dfd6f",
    "d314489b-119f-43d0-b1fa-260a6dbfb17c",
]

df_v4_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v4_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_4)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v4.csv")

## Fifth pass

REMOVED_SVIS_5 = [
    "0e2906d6-e61a-4b08-bc46-4b91b55f8e6a",
    "96ca31a0-9729-45b9-91dc-e138c543566f",
    "339db626-bf2c-4ab4-a3f4-ef7719a3998d",
    "98864503-8496-411f-a894-86042fd0d23b",
    "b12933ce-4105-4d21-b8c5-d9777bcc9d8c",
    "b9360691-eaec-4443-bb2f-f91aac116128",
    "e843b57b-6d17-4512-acbd-1ea868bc14f0",
    "e1805d35-00bd-4161-8272-39f9e7eee751",
]

df_v5_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v5_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_5)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v5.csv")

## Sixth pass

REMOVED_SVIS_6 = [
    "1d41ea4d-7d30-4284-9a29-3f15903afc21",
    "6bb34eb1-adc3-4b99-8b34-993362ab41ae",
    "d630db07-cd16-491e-aa97-fd71b7ce5ea3",
    "f1187b27-5601-49f4-b2a4-16e425bda601",
    "649fa4e4-9d44-4aa1-878f-ed3b0d5b174a",
    # extra
]

df_v6_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v6_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_5)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_amsterdam_v6.csv")

# Last QC

# final selected pool of samples
df_clustered = pd.read_csv("download_imgs/clustered_points.csv")
df_clustered.shape

# final selected images 100 images
def get_file_names_without_extension(folder_path):
    file_names = []
    for file_name in os.listdir(folder_path):
        if os.path.isfile(os.path.join(folder_path, file_name)):
            name, _ = os.path.splitext(file_name)
            file_names.append(name)
    return file_names

# Example usage
folder_path = "data/svi/sampled/Final_80/Amsterdam"
final_80 = get_file_names_without_extension(folder_path)

folder_path = "data/svi/sampled/Final_100/Amsterdam"
final_100 = get_file_names_without_extension(folder_path)

# images highlighted by co-authors
sampled = [
    "dffc7755-2e66-475b-b102-b6903e1950d2",
    "81ae8532-1729-429a-adb5-21a8c5db6fb1",
    "83ad2efd-033b-4c3a-a65c-699cd777ad83",
]

# add lists
to_remove = final_80 + final_100 + sampled + REMOVED_SVIS_1 + REMOVED_SVIS_2 + REMOVED_SVIS_3 + REMOVED_SVIS_4 + REMOVED_SVIS_5 + REMOVED_SVIS_6

# filter only this city
df_city = df_clustered[df_clustered["city_id"] == 1528355309]  # amsterdam

# remove already sampled images as well as the ones to remove
df_removed = df_city[~df_city["uuid_x"].isin(to_remove)]
df_removed.shape

sampled_from_cluster = df_removed.sample(n=200, replace=False)

sampled_from_cluster.shape

sampled_from_cluster["uuid"] = sampled_from_cluster["uuid_x"]
sampled_from_cluster.to_csv("download_imgs/filtered_points_amsterdam_final.csv")
