==> 4_resample_missing_Singapore.ipynb <==
# Steps to add new samples
1. Run entire pipeline and save images in `v1/`
2. Manually go through images and write down the name (`uuid`) of the images that are not suitable.
3. Remove this images and save the resulting ones in `v2/`
4. Launch this notebook to download missing samples. Repeat as needed

import os

import numpy as np
import pandas as pd

df_clustered = pd.read_csv("download_imgs/clustered_points.csv")
df_clustered

df_metadata_downloaded = pd.read_csv("data/svi/sampled/metadata_points.csv")
df_metadata_downloaded

# Cities

city_names = {
    1702341327: "Singapore",
    1528355309: "Amsterdam",
    1840021543: "San Francisco",
    1566342270: "Abuja",
    1152554349: "Santiago",
}

# Singapore

## First pass

REMOVED_SVIS_1 = [
    # first pass
    "0f25df72-e72d-4357-a05a-82e6da350015",
    "0ef7a0cf-1a06-4598-aca0-3eeab10cd8a6",
    "2bdae933-7b49-4355-9dc5-e29c29a4d57b",
    "3de336c8-a965-44d1-80c5-1a3d9dff5a15",
    "8e3752a5-88ea-4322-b967-3fe7cfa7ce02",
    "20f09bed-492b-4f87-8a9e-22574b93a193",
    "28d1b569-9463-4bd5-a0f5-cadb04d9bf5c",
    "28e4fdd2-5a5a-4de6-a499-3531e63ebf14",
    "037fac3b-6571-4237-811d-7b7bb3d86a79",
    "55def62f-5d0e-461b-979e-babb5f968b85",
    "59c10ca6-46ee-4c1e-8851-f03fb3f825a3",
    "75a8d933-f218-4251-9d56-c0a5e158c8c5",
    "79f087a5-c349-4090-b0ae-8dcec79ff24f",
    "89d34441-7a1a-470f-b5ea-12ad52649e26",
    "71f6fff4-4f5b-41d3-ab4b-1d8ff6fd01cf",
    "201d025f-a198-4389-9144-18d71e3c1347",
    "455a0e67-c61a-4b37-9205-8d625c60ad35",
    "647e90bd-34e5-49f1-b7be-8a9335c07438",
    "706f09b1-6da7-45e8-8823-1ddb5a3d5a54",
    "856e4b71-8dc9-4017-ba09-dcae908386a3",
    "901e5282-a4c0-4e86-94c7-3a59cbe10d16",
    "924b8925-465e-44c5-8e88-578d6ec1e8e5",
    "2168e189-4d7f-4eb7-8ebc-a6ba7db25295",
    "9404c846-bb01-4d8d-a0c5-e4717f1442e5",
    "699f4df8-b799-4a3b-bf89-9e60da3f2b17",
    "79941d56-6f59-4dfa-b9df-88b8474c149c",
    "5479868f-ae4e-420d-ac13-2aea0a5ef75d",
    "35644017-ec8a-4641-8cad-23e6e6eeb64f",
    "a74a6047-871f-4d9f-bea4-9b7ad90b755f",
    "ae297080-444a-4347-b120-34bca5dcba55",
    "aed6106e-a3c4-4cf6-bf97-d53986bc2b2d",
    "affad11e-c27f-4f75-9d35-f6464aca7306",
    "b30a45b9-76d1-4ecf-8550-12fa44bb0384",
    "b42b8248-4836-47dc-8192-a1520e424e86",
    "b44471ac-c39a-48ed-8310-06bf4a3350f4",
    "bad323b4-8e47-4577-bed1-ccae403180f4",
    "bc2e01db-1262-4dbe-b0f5-3d9bda0c3f0f",
    "c43d3e00-54bc-4ef5-a353-6ba0e0af029d",
    "ca04ed20-173d-4a13-8390-02e687f20c4e",
    "d11f1560-f10d-436f-a949-9fd18b82a84f",
    "dc723f3f-c823-445c-950a-5d2350e04e03",
    "e6b98998-c0ad-4384-a5d8-634e597abb8b",
    "f32fc6a6-b4bc-4ba1-8980-710ac487b4ba",
    "f80d7c2c-4423-45c8-b385-cd1e2425a7c9",
    "f9150b90-e918-46ef-a3c6-e4c00b9e7db7",
    "f33495f8-e8bb-4f74-a543-d14504683ab8",
    "fc395d28-7666-4910-81ef-daac56265db2",
    "fd6f7ca9-b1f8-45d5-8c70-d3f14591d002",
    "fe2c8f97-70bf-4356-b4b0-f4e1430a03c2",
]

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_1)]
df_removed.groupby("cluster_label").size()

df_city = df_clustered[df_clustered["city_id"] == 1702341327]  # singapore

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_downloaded["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_singapore.csv")

## Second pass

REMOVED_SVIS_2 = [  # second pass
    "0e5cee01-3330-452a-8033-afd90090bb6a",
    "1c80da17-697e-485b-b513-7925b07d8d92",
    "4a3697f3-027c-42da-b175-211ff2e3f927",
    "8aa713a2-8980-4522-a3d7-25de8c20efb8",
    "9aefb005-e4f5-4b2d-96d7-99f0eed4ab9e",
    "9c61e8a6-e1ce-4ef7-bf89-b245ef3f6fd0",
    "43ae2d8f-315d-4ccb-9688-fc985225ca1a",
    "47e93abf-8ca7-47f6-b83d-03ba0e8d5f5e",
    "66b59533-c6a2-4a9c-a0a0-ed27a0f6a7c6",
    "067d0122-aee7-46d2-a420-578a3e6afa11",
    "320c9508-edd0-437a-9994-2430e96238db",
    "456e1b2a-6c1d-48b7-b068-4b46ff45c877",
    "0483b263-0b6b-4709-970f-1f44b6e97b2a",
    "0608f073-3635-4432-9125-e94cf2be8a71",
    "0925ccda-5c0e-4b19-bcc3-9f661cd946f0",
    "46699e6d-5b09-44d0-be5f-1c7bcd6f8254",
    "71563d96-84c7-41f6-9565-29dad761824f",
    "98842b5b-820d-4604-ab4e-f4d2916e4b3d",
    "4783359f-1386-4bf1-901f-3e8367c9f80a",
    "05405970-be96-44ab-916d-e55eb13441de",
    "8068396d-f0ab-4475-ae62-342a0606295c",
    "89643624-40f7-4833-b12c-d38fe51384a7",
    "a520ff32-637a-4632-87bc-486899feab0c",
    "bdd16041-4586-4fa7-a5d1-e6d291829674",
    "bff0b9ee-6692-4682-9367-0193a7812355",
    "c8904169-272c-44d1-8196-9e1326de46ca",
    "cb1be2ac-c229-44c4-9b47-cc23c8fdddc2",
    "cdff144d-e8fa-4f05-bf93-2a45a845db93",
    "d17c72cf-114a-424f-b9ac-6301d1fa1310",
    "e446d921-b72a-4410-bcad-958519da7f30",
    "f1f761a9-8c10-48c4-91a4-96d4a70b3792",
    "f334b403-bb4c-4608-b9f7-548465ab73d3",
    "f829c026-4377-4e19-a03e-a9835a4ca51e",
]

df_v2_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_downloaded, df_v2_metadata])

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_2)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_singapore_v2.csv")

## Third pass

REMOVED_SVIS_3 = [
    "0b6e345f-4afd-46c9-afe2-eaf659b241dc",
    "3d4d1868-dc55-4764-a794-f851f45dc65a",
    "7a377500-89b6-4a05-8f5d-30717665a3da",
    "7c074db0-9a52-4c93-80b1-ce8768407876",
    "8c59a508-aa29-43b5-b685-08f21301da91",
    "8c443616-8da6-403f-a5bd-745f9954a29b",
    "9a063950-e248-47ca-92ca-43c546e9eea0",
    "52d37a2c-4559-4e1f-8dad-9442b0e2b8e0",
    "81a7a8f9-657a-4267-9677-73eb3601c0c9",
    "992c5755-59b4-4680-967e-58b09666eaf0",
    "48285ade-5abf-46d9-8d09-918d88639abf",
    "77201613-af62-4244-86c5-aa06898fed51",
    "b5b3a824-9cdd-45a0-8ead-527411894609",
    "b27abbc1-a94d-4fe8-a07c-6a4376995b71",
    "c8d89d9a-91a8-4916-afb6-cb419e086370",
    "ddebcdb9-37b3-459a-b98a-b60bff0e2d90",
]

df_v3_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v3_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_3)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_singapore_v3.csv")

## Fourth pass


REMOVED_SVIS_4 = [
    "3fd38e92-5f37-4027-9400-6ae5d94902d8",
    "4b88bf26-cbca-4226-b5a0-de49f85d42c0",
    "9cc2e826-46a1-4ce8-bc85-a9154a41c8b9",
    "10b9ff9c-ee42-4243-9f04-7aea70a46b05",
    "51aef8a4-9805-4a1a-a6e5-21fbfad76506",
    "89fecf54-c277-47f8-a929-724603c1d927",
    "206a63e7-5379-411b-a21d-09e6145464c8",
    "1297cc95-154a-4719-8d03-404e68dec3b5",
    "e9b5732c-8cc8-4448-a3b8-a6865d6cb958",
    "f57171dc-de01-4939-84e9-bd94b5a4ec65",
]

df_v4_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v4_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_4)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_singapore_v4.csv")

## Fifth pass

REMOVED_SVIS_5 = [
    "aafdc3db-b661-4b8b-94e9-56b10433ab04",
    "b0f8ea49-5b38-4ab0-b982-e4b6fdf8c3c2",
    "e5c4d31e-76e8-42e5-b9e7-ac7ffcd152a2",
]

df_v5_metadata = df_sampled_city[
    ["source", "orig_id", "city", "country", "continent", "uuid"]
]

df_metadata_all = pd.concat([df_metadata_all, df_v5_metadata])
df_metadata_all

# find out which cluster the removed imgs below to
df_removed = df_clustered[df_clustered["uuid_x"].isin(REMOVED_SVIS_5)]
df_removed.groupby("cluster_label").size()

# count the number of rows in each cluster that were removed
cluster_counts = df_removed["cluster_label"].value_counts()

# keep track of the samples
df_sampled_city = pd.DataFrame()

# sample the missing amount from each cluster
for cluster_label, count in cluster_counts.items():
    print(f"Sampling {count} samples from cluster {cluster_label}")

    # the available samples are the rows within this cluster that have not been already downloaded (metadata_downloaded)
    available_samples = df_city[
        (df_city["cluster_label"] == cluster_label)
        & (~df_city["uuid_x"].isin(df_metadata_all["uuid"]))
    ]

    sampled_from_cluster = available_samples.sample(n=count, replace=False)

    df_sampled_city = pd.concat([df_sampled_city, sampled_from_cluster])

df_sampled_city

df_sampled_city["uuid"] = df_sampled_city["uuid_x"]
df_sampled_city.to_csv("download_imgs/filtered_points_singapore_v5.csv")

# Last QC

# final selected pool of samples
df_clustered = pd.read_csv("download_imgs/clustered_points.csv")
df_clustered.shape

# final selected images 100 images
def get_file_names_without_extension(folder_path):
    file_names = []
    for file_name in os.listdir(folder_path):
        if os.path.isfile(os.path.join(folder_path, file_name)):
            name, _ = os.path.splitext(file_name)
            file_names.append(name)
    return file_names

# Example usage
folder_path = "data/svi/sampled/Final_100/Singapore"
final_100 = get_file_names_without_extension(folder_path)

folder_path = "data/svi/sampled/Final_80/Singapore"
final_80 = get_file_names_without_extension(folder_path)

folder_path = "data/svi/sampled/extras/Singapore"
latest = get_file_names_without_extension(folder_path)

# images highlighted by co-authors
sampled = [
    "8a467f2c-d792-4e4d-9d65-67c75f8d1f20"
]

# add lists
to_remove = final_100 + final_80 + latest + sampled + REMOVED_SVIS_1 + REMOVED_SVIS_2 + REMOVED_SVIS_3 + REMOVED_SVIS_4 + REMOVED_SVIS_5

# filter only this city
df_city = df_clustered[df_clustered["city_id"] == 1702341327]  # Singapore

# remove already sampled images as well as the ones to remove
df_removed = df_city[~df_city["uuid_x"].isin(to_remove)]
df_removed.shape

sampled_from_cluster = df_removed.sample(n=200, replace=False)
sampled_from_cluster.shape

sampled_from_cluster["uuid"] = sampled_from_cluster["uuid_x"]
sampled_from_cluster.to_csv("download_imgs/filtered_points_singapore_final.csv")
