==> 5_generate_final_svis_metadata.ipynb <==
import pandas as pd
import os

# Load list of final images

FOLDER_PATH = "data/svi/sampled/Final_80/"

# the final SVIs are in folders under the name of their respective cities
all_folders = os.listdir(FOLDER_PATH)
cities_folders = [entry for entry in all_folders if os.path.isdir(os.path.join(FOLDER_PATH, entry))]

# create empty dataframe
df_metadata = pd.DataFrame(columns=['uuid','city'])

# populate the dataframe with all uuids (filenames)
for city in cities_folders:
    # svis location
    all_SVIs = [entry for entry in os.listdir(FOLDER_PATH + city) if entry != '.DS_Store']

    # filter out only files and remove the extensions
    filenames = [os.path.splitext(entry)[0] for entry in all_SVIs if os.path.isfile(os.path.join(FOLDER_PATH + city, entry))]
    df_aux = pd.DataFrame()
    df_aux['uuid'] = filenames
    df_aux['city'] = city
    df_metadata = pd.concat([df_metadata, df_aux], ignore_index=True)


df_metadata

df_metadata.to_csv("metadata_80_final.csv")
