==> 7_personalities_q3_q1.ipynb <==
import numpy as np
import pandas as pd
from utils import (
    calculate_qscore,
    calculate_qscore_by_demographic,
)

# Load data

dataset_file = "data/labels/processed/global_mapped_cleaned.csv"
raw_df = pd.read_csv(dataset_file, index_col=0)
countries = raw_df["Country"].unique()

# Calculate Q-scores

In [PP2](https://arxiv.org/abs/1608.01769), page 6, the max comparison across all indicators are:
- min 2.40
- avg 4.05, however on page 7 they state 3.35
- max 7.67 

# Extraversion

demographic_1 = "extraversion_q3_q1"
q_scores_demographic_df = pd.DataFrame(columns=['country', demographic_1, 'Image', 'Question', 'Score', 'Num_comparisons'])
print(raw_df[demographic_1].unique())
raw_df = raw_df.dropna(subset=[demographic_1])
print(raw_df[demographic_1].unique())

# all
all_q_scores_dict = calculate_qscore_by_demographic(
        raw_df, demographic_1, min_filter=4
    )
country = 'All'
for d in all_q_scores_dict.keys():
    # transform dictionaries into a dataframe
    aux_df = all_q_scores_dict[d]
    aux_df['country'] = country
    aux_df[demographic_1] = d
    # update main df
    q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
q_scores_demographic_df

# for each country
countries_q_scores_dict = {}
for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    countries_q_scores_dict[c] = calculate_qscore_by_demographic(
        country_df, demographic_1, min_filter=4
    )

    for d in countries_q_scores_dict[c].keys():
        # transform dictionaries into a dataframe
        aux_df = countries_q_scores_dict[c][d]
        aux_df['country'] = c
        aux_df[demographic_1] = d
        # update main df
        q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
q_scores_demographic_df

# save dataframe to csv
q_scores_demographic_df.to_csv(f"data/labels/processed/{demographic_1}_qscores.csv", index=False)


# Agreeableness

demographic_1 = "agreeableness_q3_q1"
raw_df = pd.read_csv(dataset_file, index_col=0)
q_scores_demographic_df = pd.DataFrame(columns=['country', demographic_1, 'Image', 'Question', 'Score', 'Num_comparisons'])
print(raw_df[demographic_1].unique())
raw_df = raw_df.dropna(subset=[demographic_1])
print(raw_df[demographic_1].unique())

# all
all_q_scores_dict = calculate_qscore_by_demographic(
        raw_df, demographic_1, min_filter=4
    )
country = 'All'

for d in all_q_scores_dict.keys():
    # transform dictionaries into a dataframe
    aux_df = all_q_scores_dict[d]
    aux_df['country'] = country
    aux_df[demographic_1] = d
    # update main df
    q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# for each country
countries_q_scores_dict = {}
for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    countries_q_scores_dict[c] = calculate_qscore_by_demographic(
        country_df, demographic_1, min_filter=4 
    )

    for d in countries_q_scores_dict[c].keys():
        # transform dictionaries into a dataframe
        aux_df = countries_q_scores_dict[c][d]
        aux_df['country'] = c
        aux_df[demographic_1] = d
        # update main df
        q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])

q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# save dataframe to csv
q_scores_demographic_df.to_csv(f"data/labels/processed/{demographic_1}_qscores.csv", index=False)


# Conscientiousness

demographic_1 = "conscientiousness_q3_q1"
raw_df = pd.read_csv(dataset_file, index_col=0)
q_scores_demographic_df = pd.DataFrame(columns=['country', demographic_1, 'Image', 'Question', 'Score', 'Num_comparisons'])
print(raw_df[demographic_1].unique())
raw_df = raw_df.dropna(subset=[demographic_1])
print(raw_df[demographic_1].unique())

# all
all_q_scores_dict = calculate_qscore_by_demographic(
        raw_df, demographic_1, min_filter=4
    )
country = 'All'

for d in all_q_scores_dict.keys():
    # transform dictionaries into a dataframe
    aux_df = all_q_scores_dict[d]
    aux_df['country'] = country
    aux_df[demographic_1] = d
    # update main df
    q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
    
q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# for each country
countries_q_scores_dict = {}
for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    countries_q_scores_dict[c] = calculate_qscore_by_demographic(
        country_df, demographic_1, min_filter=4 
    )

    for d in countries_q_scores_dict[c].keys():
        # transform dictionaries into a dataframe
        aux_df = countries_q_scores_dict[c][d]
        aux_df['country'] = c
        aux_df[demographic_1] = d
        # update main df
        q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])

q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# save dataframe to csv
q_scores_demographic_df.to_csv(f"data/labels/processed/{demographic_1}_qscores.csv", index=False)


# Neuroticism

demographic_1 = "neuroticism_q3_q1"
raw_df = pd.read_csv(dataset_file, index_col=0)
q_scores_demographic_df = pd.DataFrame(columns=['country', demographic_1, 'Image', 'Question', 'Score', 'Num_comparisons'])
print(raw_df[demographic_1].unique())
raw_df = raw_df.dropna(subset=[demographic_1])
print(raw_df[demographic_1].unique())

# all
all_q_scores_dict = calculate_qscore_by_demographic(
        raw_df, demographic_1, min_filter=4
    )
country = 'All'

for d in all_q_scores_dict.keys():
    # transform dictionaries into a dataframe
    aux_df = all_q_scores_dict[d]
    aux_df['country'] = country
    aux_df[demographic_1] = d
    # update main df
    q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# for each country
countries_q_scores_dict = {}
for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    countries_q_scores_dict[c] = calculate_qscore_by_demographic(
        country_df, demographic_1, min_filter=4 
    )

    for d in countries_q_scores_dict[c].keys():
        # transform dictionaries into a dataframe
        aux_df = countries_q_scores_dict[c][d]
        aux_df['country'] = c
        aux_df[demographic_1] = d
        # update main df
        q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])

q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# save dataframe to csv
q_scores_demographic_df.to_csv(f"data/labels/processed/{demographic_1}_qscores.csv", index=False)

# Openness

demographic_1 = "openness_q3_q1"
raw_df = pd.read_csv(dataset_file, index_col=0)
q_scores_demographic_df = pd.DataFrame(columns=['country', demographic_1, 'Image', 'Question', 'Score', 'Num_comparisons'])
print(raw_df[demographic_1].unique())
raw_df = raw_df.dropna(subset=[demographic_1])
print(raw_df[demographic_1].unique())

# all
all_q_scores_dict = calculate_qscore_by_demographic(
        raw_df, demographic_1, min_filter=4
    )
country = 'All'

for d in all_q_scores_dict.keys():
    # transform dictionaries into a dataframe
    aux_df = all_q_scores_dict[d]
    aux_df['country'] = country
    aux_df[demographic_1] = d
    # update main df
    q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])
q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# for each country
countries_q_scores_dict = {}
for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    countries_q_scores_dict[c] = calculate_qscore_by_demographic(
        country_df, demographic_1, min_filter=4 
    )

    for d in countries_q_scores_dict[c].keys():
        # transform dictionaries into a dataframe
        aux_df = countries_q_scores_dict[c][d]
        aux_df['country'] = c
        aux_df[demographic_1] = d
        # update main df
        q_scores_demographic_df = pd.concat([q_scores_demographic_df, aux_df])

q_scores_demographic_df.groupby(["country", "Question", demographic_1])["Image"].nunique().reset_index()

# save dataframe to csv
q_scores_demographic_df.to_csv(f"data/labels/processed/{demographic_1}_qscores.csv", index=False)
