==> 9c_perceptual-variations.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

from utils import calculate_qscore

# Load data

dataset_file = "data/labels/processed/global_mapped_cleaned.csv"  # full dataset
metadata_file = "data/svi/metadata.csv"

raw_df = pd.read_csv(dataset_file, index_col=0)
metadata_df = pd.read_csv(metadata_file, index_col=0)
metadata_df

# Compute Q-score

Come up with a dataframe that have the scores for images based on the location of the SVI and location of participants

In [PP2](https://arxiv.org/abs/1608.01769), page 6, the max comparison across all indicators are:
- min 2.40
- avg 4.05, however on page 7 they state 3.35
- max 7.67 

to_from_df = pd.DataFrame()  # placeholder to keep track of data
min_filter = 4
countries = raw_df["Country"].unique()

for c in countries:
    country_df = raw_df[raw_df["Country"] == c]
    country_df_scores = calculate_qscore(country_df)
    country_df_scores_merged = pd.merge(
        country_df_scores,
        metadata_df,
        left_on="Image",
        right_on="Image number",
        how="left",
    )
    country_df_scores_merged["participants_from"] = c
    country_df_scores_merged["SVI_from"] = country_df_scores_merged["city"].map(
        {
            "Singapore": "Singapore",
            "San Francisco": "USA",
            "Amsterdam": "Netherlands",
            "Santiago": "Chile",
            "Abuja": "Nigeria",
        }
    )
    to_from_df = pd.concat([to_from_df, country_df_scores_merged])

to_from_df['Score'] = round(to_from_df['Score'], 2)
to_from_df.to_csv('data/labels/processed/to_from_scores.csv')

print(to_from_df['Num_comparisons'].min())
print(to_from_df['Num_comparisons'].max())
